Split research_ai models into one file per model
Convert research_ai/models.py into a models/ package with one module
per model, re-exported from __init__.py so existing imports are
unchanged. No model or schema changes.
