+++ src/research_ai/models/__init__.py
@@ -1,0 +1,15 @@
+from .email_template import EmailTemplate
+from .expert import Expert
+from .expert_search import ExpertSearch
+from .generated_email import GeneratedEmail
+from .proposal_draft import ProposalDraft
+from .search_expert import SearchExpert
+
+__all__ = [
+    "EmailTemplate",
+    "Expert",
+    "ExpertSearch",
+    "GeneratedEmail",
+    "ProposalDraft",
+    "SearchExpert",
+]
